+++ app/src/main/java/com/zybooks/irreduciblerepresentationscalculator/PointGroups/Cn_Group/c4.java
@@ -55,10 +55,10 @@ public class c4 extends AppCompatActivity {
             reset.setVisibility(View.VISIBLE);
 
             TableData td = new TableData(getApplicationContext(), "c4");
-            double[] input = new double[3];
-            input[0] = Double.parseDouble(value1.getText().toString().trim());
-            input[1] = Double.parseDouble(value2.getText().toString().trim());
-            input[2] = Double.parseDouble(value3.getText().toString().trim());
+            int[] input = new int[3];
+            input[0] = Integer.parseInt(value1.getText().toString().trim());
+            input[1] = Integer.parseInt(value2.getText().toString().trim());
+            input[2] = Integer.parseInt(value3.getText().toString().trim());
 
             hideKeyboard(view);
             td.calculate(input);
